requirement.txt
opencv-python
torch
torchvision
pandas
numpy
Pillow
colour
dlib
